make cell module
make cell module
& additional module to make cell
(common, color, console, cursor, point)
