+++ cell.h
@@ -1,9 +1,11 @@
 #ifndef __CELL_H__
 #define __CELL_H__
 
+#include "common.h"
 #include "color.h"
 #include "point.h"
 #include "blocks.h"
+#include "console.h"
 
 typedef struct _cell_t {
     point_t point; // cellÀÇ À§Ä¡
