+++ common.c
@@ -1,0 +1,1 @@
+#include "common.h"
