+++ common.h
@@ -1,6 +1,18 @@
 #ifndef __COMMON_H__
 #define __COMMON_H__
 
+#include <stdio.h>
+#include <stdint.h>
+#include <stdbool.h>
+#include <Windows.h>
+
+
+
+#define DEFAULT_COLS 120
+#define DEFAULT_ROWS 30
+
+#define RESIZE_COLS (DEFAULT_COLS-30) //=90
+#define RESIZE_ROWS 30
