+++ console.c
@@ -1,0 +1,23 @@
+#include "console.h"
+
+//console
+void console_set_fore_color(windows_console_t* console, uint8_t fore_color) {
+    console->fore_color = fore_color;
+    SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), fore_color);
+}
+//console
+void console_set_back_color(windows_console_t* console, uint8_t back_color) {
+    console->back_color = back_color;
+    SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), back_color);
+}
+//console
+void console_set_color(windows_console_t* console, uint8_t fore_color, uint8_t back_color) {
+    console->fore_color = fore_color;
+    console->back_color = back_color;
+    //void console_set_color(unsigned short text, unsigned short back) {
+    SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), console->fore_color | (console->back_color << 4));
+}
+//console
+void console_set_default_color(windows_console_t* console) {
+    console_set_color(console, WHITE, BLACK);
+}
