+++ main.c
@@ -4,84 +4,19 @@
 #include <stdbool.h>
 #include <windows.h>
 //test
+#include "common.h"
 #include "blocks.h"
 #include "key_driver.h"
+#include "console.h"
+#include "color.h"
+#include "point.h"
+#include "cell.h"
+
 
 #define _CRT_SECURE_NO_WARNINGS
 
-#define DEFAULT_COLS 120
-#define DEFAULT_ROWS 30
-
-#define RESIZE_COLS (DEFAULT_COLS-30) //=90
-#define RESIZE_ROWS 30
-
 //geunu start edit
-// 
-//추가적으로 밝은 색상은 해당 값에 8을 더한 값으로 표현됩니다(예: 밝은 파랑은 9).
-typedef enum _color_t {
-    BLACK, //0
-    BLUE, // 1
-    GREEN,
-    CYAN,
-    RED,
-    PURPLE,
-    YELLOW,
-    WHITE,
-    GRAY,
-    LIGHT_BLUE,
-    LIGHT_GREEN,
-    LIGHT_CYAN,
-    LIGHT_RED,
-    LIGHT_PURBPLE,
-    LIGHT_YELLOW,
-    LIGHT_WHITE,
-} color_t;
 
-
-//추가적으로 밝은 색상은 해당 값에 8을 더한 값으로 표현됩니다(예: 밝은 파랑은 9).
-char* color_strings[16] = {
-    "black", // 검정
-    "blue", // 파랑
-    "green", // 초록
-    "cyan", // 청록
-    "red", // 빨강
-    "purple", // 자주
-    "yellow", // 노랑
-    "white", // 흰색
-    "gray", // 회색
-    "light blue", // 밝은 파랑
-    "light green", // 밝은 회색
-    "light cyan", // 밝은 청록
-    "light red", // 밝은 빨강
-    "light purple", // 밝은 자주
-    "light yellow", // 밝은 노랑
-    "light white" // 밝은 흰색
-};
-
-char* get_color_string(color_t color) {
-    if (color >= BLACK && color <= LIGHT_WHITE) {
-        return color_strings[color]; // 유효한 색상 값이면 해당 문자열 반환
-    }
-    else {
-        return "unknown"; // 유효하지 않은 값이면 기본값 반환
-    }
-}
-
-typedef struct _cursor_info_t {
-    uint32_t x;
-    uint32_t y;
-    bool is_hide;
-} cursor_info_t;
-
-typedef struct _windows_console_t {
-    uint32_t cols;
-    uint32_t rows;
-
-    color_t fore_color;
-    color_t back_color;
-
-    cursor_info_t cursor_info;
-} windows_console_t;
 //console
 void console_resize(const windows_console_t* console, uint32_t cols, uint32_t rows);
 
@@ -203,45 +138,10 @@ void console_display_info(const windows_console_t* console) {
     printf("[%d] Colors (fore,back) = (%s,%s)\r\n", ++dummy_count, get_color_string(console->fore_color), get_color_string(console->back_color));
     printf("[%d] Cursor (x,y,hide) = %d,%d,%d)\r\n", ++dummy_count, console->cursor_info.x, console->cursor_info.y, console->cursor_info.is_hide);
 }
-//point
-typedef struct _point_t {
-    uint32_t x;
-    uint32_t y;
-} point_t;
-//cell
-typedef struct _cell_t {
-    point_t point; // cell의 위치
-    color_t color; // cell의 색상
-    cell_attributes_t att; // cell의 속성
-} cell_t;
-//cell
-void cell_init(cell_t* cell, int x, int y, cell_attributes_t att, color_t color) {
-    cell->point.x = x;
-    cell->point.y = y;
-    cell->att = att;
-    cell->color = color;
-}
-//console
-void console_set_fore_color(windows_console_t* console, uint8_t fore_color) {
-    console->fore_color = fore_color;
-    SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), fore_color);
-}
-//console
-void console_set_back_color(windows_console_t* console, uint8_t back_color) {
-    console->back_color = back_color;
-    SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), back_color);
-}
-//console
-void console_set_color(windows_console_t* console, uint8_t fore_color, uint8_t back_color) {
-    console->fore_color = fore_color;
-    console->back_color = back_color;
-    //void console_set_color(unsigned short text, unsigned short back) {
-    SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), console->fore_color | (console->back_color << 4));
-}
-//console
-void console_set_default_color(windows_console_t* console) {
-    console_set_color(console, WHITE, BLACK);
-}
+
+
+
+
 //console
 void console_set_fill_color(int background_color) {
     HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE); // 콘솔 핸들 가져오기
@@ -332,21 +232,7 @@ void console_clear(const windows_console_t* console) {
 void setcolor(unsigned short text, unsigned short back) {
     SetConsoleTextAttribute(GetStdHandle(STD_OUTPUT_HANDLE), text | (back << 4));
 }
-//cell
-//void draw_one_cell(cell_t* cell, uint32_t x, uint32_t y) {
-void cell_draw(cell_t* cell, uint32_t x, uint32_t y) {
-    cell->point.x = x;
-    cell->point.y = y;
 
-    COORD pos;
-    pos.X = (short)cell->point.x;
-    pos.Y = (short)cell->point.y;
-    pos.X *= 2;
-
-    HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE);
-    SetConsoleCursorPosition(hConsole, pos);
-    printf("■");
-}
 
 //block
 void block_draw(block_t* block) {
@@ -400,18 +286,7 @@ void block_draw(const block_t* blk, uint32_t x, uint32_t y) {
 }
 */
 
-//cell
-void cell_draw_test(windows_console_t* console) {
-    cell_t my_cell;
-    cell_init(&my_cell, 0, 0, N, YELLOW); // 좌표계 점검용 테스트 코드
-    console_set_fore_color(console, my_cell.color);
-    //console->fore_color(my_cell.color);
-    cell_draw(&my_cell, 0, 0); // 좌상단
-    cell_draw(&my_cell, RESIZE_COLS / 2 - 1, 0); // 우상단
-    cell_draw(&my_cell, RESIZE_COLS / 2 - 1, RESIZE_ROWS - 1); // 우하단
-    cell_draw(&my_cell, 0, RESIZE_ROWS - 1); // 좌하단
-    cell_draw(&my_cell, (RESIZE_COLS / 2 - 1) / 2, (RESIZE_ROWS - 1) / 2); // 정중앙
-}
+
 
 bool running = true;
