+++ tetris_r4.vcxproj
@@ -128,14 +128,26 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="blocks.c" />
+    <ClCompile Include="cell.c" />
+    <ClCompile Include="color.c" />
+    <ClCompile Include="common.c" />
+    <ClCompile Include="console.c" />
+    <ClCompile Include="cursor.c" />
     <ClCompile Include="key_driver.c" />
     <ClCompile Include="main.c" />
+    <ClCompile Include="point.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="blocks.h" />
+    <ClInclude Include="cell.h" />
+    <ClInclude Include="color.h" />
+    <ClInclude Include="common.h" />
+    <ClInclude Include="console.h" />
+    <ClInclude Include="cursor.h" />
     <ClInclude Include="key_code.h" />
     <ClInclude Include="key_config.h" />
     <ClInclude Include="key_driver.h" />
+    <ClInclude Include="point.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
